requirements.txt
# scraper
bs4
# api

# llm
langchain==0.0.207
openai==0.27.8
kor==0.10.0
chromadb==0.3.26
python-dotenv  
